feat: cập nhật hệ thống định vị thời gian thực, module phân tích dữ liệu và hoàn thiện tài liệu tiểu luận final

+++ ggmap_user/app/src/main/java/com/example/catagentdeployer/network/ApiService.kt
@@ -21,7 +21,7 @@ data class Place(
 
 data class PlacesResponse(val places: List<Place>, val total: Int)
 
-data class DirectionDistance(val text: String, val value: Int)
+data class DirectionDistance(val text: String, val value: Double)
 data class DirectionsResponse(
     val distance:         DirectionDistance?,
     val duration:         DirectionDistance?,
